refactor(internal/resources): consolidate 404 checking into errors.IsNotFound
Replace inline *fastly.HTTPError type assertions with wrapped-error-safe
  errors.IsNotFound helper. Move implementation from internal/client to new
  internal/errors package to avoid import cycles.

  Changes:
  - Create internal/errors/http.go with canonical IsNotFound implementation using errors.As
  - Remove internal/client/errors.go (was forwarding wrapper)
  - Delete private service.isNotFound function (used bare type assertion)
  - Update 18 call sites across resources, service, computepackage, and tests to use errors.IsNotFound
