+++ internal/errors/http.go
@@ -1,4 +1,4 @@
-package client
+package errors
 
 import (
 	"errors"
