+++ internal/resources/servicecdnauto/resource.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	fastlyclient "github.com/fastly/terraform-provider-fastly/internal/client"
+	"github.com/fastly/terraform-provider-fastly/internal/errors"
 	"github.com/fastly/terraform-provider-fastly/internal/resources/backend"
 	"github.com/fastly/terraform-provider-fastly/internal/resources/domain"
 	"github.com/fastly/terraform-provider-fastly/internal/service"
@@ -188,7 +189,7 @@ func (r *Resource) Read(ctx context.Context, req resource.ReadRequest, resp *res
 		ServiceID: state.ID.ValueString(),
 	})
 	if err != nil {
-		if httpErr, ok := err.(*fastly.HTTPError); ok && httpErr.StatusCode == 404 {
+		if errors.IsNotFound(err) {
 			resp.State.RemoveResource(ctx)
 			return
 		}
